Rework Latex test for nbconvert Python API
Also, skip it if pandoc is not available
